Stop allowing identifying relationships from overriding cascade behavior
Fixes #28961

In the issue, an alternate key was added over the the foreign key to make it effectively an identifying relationship. That is, changing the relationship changes the identity of the dependent. This resulted in the special handling for identifying relationships to kick in, which didn't respect the configured DeleteBehavior. This change fixes that.

+++ src/EFCore/ChangeTracking/Internal/NavigationFixer.cs
@@ -1416,28 +1416,39 @@ public class NavigationFixer : INavigationFixer
 
         if (foreignKey.IsRequired
             && hasOnlyKeyProperties
-            && dependentEntry.EntityState != EntityState.Detached)
+            && dependentEntry.EntityState != EntityState.Detached
+            && dependentEntry.EntityState != EntityState.Deleted)
         {
-            try
+            if (foreignKey.DeleteBehavior == DeleteBehavior.Cascade
+                || foreignKey.DeleteBehavior == DeleteBehavior.ClientCascade
+                || foreignKey.IsOwnership)
             {
-                _inFixup = true;
-                switch (dependentEntry.EntityState)
+                try
                 {
-                    case EntityState.Added:
-                        dependentEntry.SetEntityState(EntityState.Detached);
-                        DeleteFixup(dependentEntry);
-                        break;
-                    case EntityState.Unchanged:
-                    case EntityState.Modified:
-                        dependentEntry.SetEntityState(
-                            dependentEntry.SharedIdentityEntry != null ? EntityState.Detached : EntityState.Deleted);
-                        DeleteFixup(dependentEntry);
-                        break;
+                    _inFixup = true;
+                    switch (dependentEntry.EntityState)
+                    {
+                        case EntityState.Added:
+                            dependentEntry.SetEntityState(EntityState.Detached);
+                            DeleteFixup(dependentEntry);
+                            break;
+                        case EntityState.Unchanged:
+                        case EntityState.Modified:
+                            dependentEntry.SetEntityState(
+                                dependentEntry.SharedIdentityEntry != null ? EntityState.Detached : EntityState.Deleted);
+                            DeleteFixup(dependentEntry);
+                            break;
+                    }
+                }
+                finally
+                {
+                    _inFixup = false;
                 }
             }
-            finally
+            else
             {
-                _inFixup = false;
+                throw new InvalidOperationException(
+                    CoreStrings.KeyReadOnly(dependentProperties.First().Name, dependentEntry.EntityType.DisplayName()));
             }
         }
     }
